updating models and moving games to their own
